requirements.txt
requests==2.22.0
opencensus==0.7.6    
opencensus-context==0.1.1    
opencensus-ext-azure==1.0.1 
opencensus-ext-flask==0.7.3 
